HSEARCH-2632 Use only one job-level listener
The execution order of multiple job-level listeners are not sure, so
it's preferable to use only one listener.
